requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
mysql-connector-python==8.1.0
PyMySQL==1.1.0
Werkzeug==2.3.7
pandas==2.1.1
scikit-learn==1.5.1
qrcode==7.4.2
reportlab==4.0.4
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.26.4
faker==19.6.2
Flask-WTF==1.2.1
WTForms==3.1.2
gunicorn==21.2.0
